+++ src/main/java/com/home_rental_solution/ms_inquilinos/model/Inquilino.java
@@ -17,7 +17,7 @@ public class Inquilino {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idInquilino;
+    private Long idInquilino;
 
     @Column(nullable = false, length = 120)
     private String nombre;
